Update README.md
Input Function in Python

The input function is an essential feature in Python that allows to take input from the user. This is particularly useful when you want to create interactive programs where the user can provide data during execution. Also known as user input function.

How input Function Works:

The input function waits for the user to type something and then press Enter. It reads the input as a string and returns it.

Example: 
#Prompting the user for their name
Name=input("Enter your name: ")